pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "{{PackageName}}"
version = "0.9.2"
description = "A plugin for django-simple-deploy, supporting deployments to {{PlatformName}}."
readme = "README.md"

authors = [
    {name = "Eric Matthes", email = "ehmatthes@gmail.com" },
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

keywords = ["django", "deployment"]

requires-python = ">=3.9"

dependencies = [
    "django>=4.2",
    "pluggy>=1.5.0",
    "toml>=0.10.2",
    "requests>=2.32.2",
    "django-simple-deploy>=0.9.0"
]

[project.optional-dependencies]
dev = [
    "black>=24.1.0",
    "build>=1.2.1",
    "pytest>=8.3.0",
    "twine>=5.1.1",
]

# [project.urls]
# "Documentation" = "<documentation URL>"
# "GitHub" = "<GH repo URL>"
# "Changelog" = "<changelog URL"

[tool.setuptools]
packages = [
    "{{PluginName}}",
    "{{PluginName}}.templates",
]
include-package-data = true
